Refactor Debtors note to follow the new pattern
