--- src/test/java/uk/gov/companieshouse/api/accounts/controller/DebtorsControllerTest.java
@@ -1,298 +1,0 @@
-package uk.gov.companieshouse.api.accounts.controller;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.when;
-
-import java.util.Map;
-import javax.servlet.http.HttpServletRequest;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
-import uk.gov.companieshouse.api.accounts.AttributeName;
-import uk.gov.companieshouse.api.accounts.exception.DataException;
-import uk.gov.companieshouse.api.accounts.links.SmallFullLinkType;
-import uk.gov.companieshouse.api.accounts.model.rest.SmallFull;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.debtors.Debtors;
-import uk.gov.companieshouse.api.accounts.model.validation.Errors;
-import uk.gov.companieshouse.api.accounts.service.impl.DebtorsService;
-import uk.gov.companieshouse.api.accounts.service.response.ResponseObject;
-import uk.gov.companieshouse.api.accounts.service.response.ResponseStatus;
-import uk.gov.companieshouse.api.model.transaction.Transaction;
-import uk.gov.companieshouse.api.accounts.utility.ApiResponseMapper;
-import uk.gov.companieshouse.api.accounts.utility.ErrorMapper;
-
-@ExtendWith(MockitoExtension.class)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class DebtorsControllerTest {
-
-    private static final String COMPANY_ACCOUNTS_ID = "companyAccountsId";
-
-    @Mock
-    private BindingResult mockBindingResult;
-
-    @Mock
-    private Debtors mockDebtors;
-
-    @Mock
-    private Transaction mockTransaction;
-
-    @Mock
-    private HttpServletRequest mockRequest;
-
-    @Mock
-    private ApiResponseMapper mockApiResponseMapper;
-
-    @Mock
-    private DebtorsService mockDebtorsService;
-
-    @Mock
-    private SmallFull mockSmallFull;
-
-    @Mock
-    private Map<String, String> mockSmallFullLinks;
-
-    @Mock
-    private ErrorMapper mockErrorMapper;
-
-    @InjectMocks
-    DebtorsController controller;
-
-    @Test
-    @DisplayName("debtors resource created successfully")
-    void createDebtorsResource() throws DataException {
-
-        when(mockBindingResult.hasErrors()).thenReturn(false);
-
-        when(mockRequest.getAttribute(AttributeName.TRANSACTION.getValue())).thenReturn(mockTransaction);
-
-        ResponseObject responseObject = new ResponseObject(ResponseStatus.CREATED,
-            mockDebtors);
-        when(mockDebtorsService.create(mockDebtors, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest))
-            .thenReturn(responseObject);
-
-        ResponseEntity responseEntity = ResponseEntity.status(HttpStatus.CREATED)
-            .body(responseObject.getData());
-        when(mockApiResponseMapper.map(responseObject.getStatus(), responseObject.getData(), responseObject.getErrors()))
-            .thenReturn(responseEntity);
-
-        ResponseEntity returnedResponse =
-            controller.create(mockDebtors, mockBindingResult, COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(returnedResponse);
-        assertEquals(HttpStatus.CREATED, responseEntity.getStatusCode());
-        assertEquals(mockDebtors, responseEntity.getBody());
-    }
-
-    @Test
-    @DisplayName("Create debtors has failed - data exception thrown")
-    void createDebtorsDataException() throws DataException {
-
-        when(mockBindingResult.hasErrors()).thenReturn(false);
-        when(mockRequest.getAttribute(AttributeName.TRANSACTION.getValue())).thenReturn(mockTransaction);
-
-        when(mockDebtorsService.create(mockDebtors, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest))
-            .thenThrow(new DataException(""));
-
-        ResponseEntity responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        when(mockApiResponseMapper.getErrorResponse())
-            .thenReturn(responseEntity);
-
-        ResponseEntity returnedResponse =
-            controller.create(mockDebtors, mockBindingResult, COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(returnedResponse);
-        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-        assertNull(responseEntity.getBody());
-    }
-
-    @Test
-    @DisplayName("Create debtors - has binding errors")
-    void createDebtorsBindingErrors() {
-
-        when(mockBindingResult.hasErrors()).thenReturn(true);
-        when(mockErrorMapper.mapBindingResultErrorsToErrorModel(mockBindingResult)).thenReturn(new Errors());
-
-        ResponseEntity responseEntity =
-            controller.create(mockDebtors, mockBindingResult, COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(responseEntity);
-        assertEquals(HttpStatus.BAD_REQUEST, responseEntity.getStatusCode());
-    }
-
-    @Test
-    @DisplayName("Get debtors - success")
-    void getDebtorsSuccess() throws DataException {
-
-        when(mockRequest.getAttribute(AttributeName.TRANSACTION.getValue())).thenReturn(mockTransaction);
-
-        ResponseObject responseObject = new ResponseObject(ResponseStatus.FOUND,
-            mockDebtors);
-        when(mockDebtorsService.find(COMPANY_ACCOUNTS_ID, mockRequest))
-            .thenReturn(responseObject);
-
-        ResponseEntity responseEntity = ResponseEntity.status(HttpStatus.FOUND)
-            .body(responseObject.getData());
-        when(mockApiResponseMapper.mapGetResponse(responseObject.getData(), mockRequest))
-            .thenReturn(responseEntity);
-
-        ResponseEntity returnedResponse =
-            controller.get(COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(returnedResponse);
-        assertEquals(HttpStatus.FOUND, responseEntity.getStatusCode());
-        assertEquals(mockDebtors, responseEntity.getBody());
-    }
-
-    @Test
-    @DisplayName("Get debtors - data exception thrown")
-    void getDebtorsDataException() throws DataException {
-
-        when(mockRequest.getAttribute(AttributeName.TRANSACTION.getValue())).thenReturn(mockTransaction);
-
-        when(mockDebtorsService.find(COMPANY_ACCOUNTS_ID, mockRequest))
-            .thenThrow(new DataException(""));
-
-        ResponseEntity responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        when(mockApiResponseMapper.getErrorResponse()).thenReturn(responseEntity);
-
-        ResponseEntity returnedResponse = controller.get(COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(returnedResponse);
-        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-        assertNull(responseEntity.getBody());
-    }
-
-    @Test
-    @DisplayName("Update debtors - no small full link")
-    void updateDebtorsNoSmallFullLink() {
-
-        when(mockRequest.getAttribute(AttributeName.SMALLFULL.getValue())).thenReturn(mockSmallFull);
-        when(mockSmallFull.getLinks()).thenReturn(mockSmallFullLinks);
-        when(mockSmallFullLinks.get(SmallFullLinkType.DEBTORS_NOTE.getLink())).thenReturn(null);
-
-        ResponseEntity responseEntity =
-            controller.update(mockDebtors, mockBindingResult, COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(responseEntity);
-        assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode());
-        assertNull(responseEntity.getBody());
-    }
-
-    @Test
-    @DisplayName("Update debtors - has binding errors")
-    void updateDebtorsBindingErrors() {
-
-        mockTransactionAndLinks();
-        when(mockBindingResult.hasErrors()).thenReturn(true);
-        when(mockErrorMapper.mapBindingResultErrorsToErrorModel(mockBindingResult)).thenReturn(new Errors());
-
-        ResponseEntity responseEntity =
-            controller.update(mockDebtors, mockBindingResult, COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(responseEntity);
-        assertEquals(HttpStatus.BAD_REQUEST, responseEntity.getStatusCode());
-        assertNotNull(responseEntity.getBody());
-    }
-
-    @Test
-    @DisplayName("Update debtors - success")
-    void updateDebtorsSuccess() throws DataException {
-
-        mockTransactionAndLinks();
-        when(mockBindingResult.hasErrors()).thenReturn(false);
-
-        ResponseObject responseObject = new ResponseObject(ResponseStatus.UPDATED,
-            mockDebtors);
-        when(mockDebtorsService.update(mockDebtors, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest))
-            .thenReturn(responseObject);
-
-        ResponseEntity responseEntity = ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-        when(mockApiResponseMapper.map(responseObject.getStatus(), responseObject.getData(), responseObject.getErrors()))
-            .thenReturn(responseEntity);
-
-        ResponseEntity returnedResponse =
-            controller.update(mockDebtors, mockBindingResult, COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(returnedResponse);
-        assertEquals(HttpStatus.NO_CONTENT, responseEntity.getStatusCode());
-        assertNull(responseEntity.getBody());
-    }
-
-    @Test
-    @DisplayName("Update debtors - data exception thrown")
-    void updateDebtorsDataException() throws DataException {
-
-        mockTransactionAndLinks();
-        when(mockBindingResult.hasErrors()).thenReturn(false);
-
-        when(mockDebtorsService.update(mockDebtors, mockTransaction, COMPANY_ACCOUNTS_ID, mockRequest))
-            .thenThrow(new DataException(""));
-
-        ResponseEntity responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        when(mockApiResponseMapper.getErrorResponse()).thenReturn(responseEntity);
-
-        ResponseEntity returnedResponse =
-            controller.update(mockDebtors, mockBindingResult, COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(returnedResponse);
-        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-        assertNull(responseEntity.getBody());
-    }
-
-    @Test
-    @DisplayName("Delete debtors - success")
-    void deleteDebtorsSuccess() throws DataException {
-
-        when(mockRequest.getAttribute(AttributeName.TRANSACTION.getValue())).thenReturn(mockTransaction);
-
-        ResponseObject responseObject = new ResponseObject(ResponseStatus.UPDATED,
-            mockDebtors);
-        when(mockDebtorsService.delete(COMPANY_ACCOUNTS_ID, mockRequest))
-            .thenReturn(responseObject);
-
-        ResponseEntity responseEntity = ResponseEntity.status(HttpStatus.NO_CONTENT)
-            .build();
-        when(mockApiResponseMapper.map(responseObject.getStatus(), responseObject.getData(), responseObject.getErrors()))
-            .thenReturn(responseEntity);
-
-        ResponseEntity returnedResponse =
-            controller.delete(COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(returnedResponse);
-        assertEquals(HttpStatus.NO_CONTENT, responseEntity.getStatusCode());
-    }
-
-    @Test
-    @DisplayName("Delete debtors - data exception thrown")
-    void deleteDebtorsDataException() throws DataException {
-
-        when(mockRequest.getAttribute(AttributeName.TRANSACTION.getValue())).thenReturn(mockTransaction);
-
-        when(mockDebtorsService.delete(COMPANY_ACCOUNTS_ID, mockRequest))
-            .thenThrow(new DataException(""));
-
-        ResponseEntity responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        when(mockApiResponseMapper.getErrorResponse()).thenReturn(responseEntity);
-
-        ResponseEntity returnedResponse = controller.delete(COMPANY_ACCOUNTS_ID, mockRequest);
-
-        assertNotNull(returnedResponse);
-        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, responseEntity.getStatusCode());
-    }
-
-    private void mockTransactionAndLinks() {
-        when(mockRequest.getAttribute(anyString())).thenReturn(mockSmallFull).thenReturn(mockTransaction);
-        when(mockSmallFull.getLinks()).thenReturn(mockSmallFullLinks);
-        when(mockSmallFullLinks.get(SmallFullLinkType.DEBTORS_NOTE.getLink())).thenReturn("");
-    }
-}
